+++ src/phases/phase_two/stages/stage_two/prompts/abstract_prompts.py
