+++ src/phases/phase_two/stages/stage_two/prompts/outline_prompts.py
