+++ src/phases/phase_two/stages/stage_two/workers/abstract_worker.py
