+++ src/phases/phase_two/stages/stage_two/workers/outline_critic.py
