+++ src/phases/phase_two/stages/stage_two/workers/outline_refinement_worker.py
