+++ tests/test_key_moves_cycles.py
@@ -4,9 +4,9 @@ import json
 import yaml
 import os
 from pathlib import Path
-from src.stages.phase_two.stages.stage_two.workers.key_moves_worker import KeyMovesWorker
-from src.stages.phase_two.stages.stage_two.workers.key_moves_critic import KeyMovesCritic
-from src.stages.phase_two.stages.stage_two.workers.key_moves_refinement_worker import KeyMovesRefinementWorker
+from src.phases.phase_two.stages.stage_two.workers.key_moves_worker import KeyMovesWorker
+from src.phases.phase_two.stages.stage_two.workers.key_moves_critic import KeyMovesCritic
+from src.phases.phase_two.stages.stage_two.workers.key_moves_refinement_worker import KeyMovesRefinementWorker
 
 def load_yaml(path: str) -> dict:
     """Load YAML file"""
