+++ tests/test_outline_cycle.py
@@ -4,9 +4,9 @@ import json
 import yaml
 import os
 from pathlib import Path
-from src.stages.phase_two.stages.stage_two.workers.outline_worker import OutlineWorker
-from src.stages.phase_two.stages.stage_two.workers.outline_critic import OutlineCritic
-from src.stages.phase_two.stages.stage_two.workers.outline_refinement_worker import OutlineRefinementWorker
+from src.phases.phase_two.stages.stage_two.workers.outline_worker import OutlineWorker
+from src.phases.phase_two.stages.stage_two.workers.outline_critic import OutlineCritic
+from src.phases.phase_two.stages.stage_two.workers.outline_refinement_worker import OutlineRefinementWorker
 
 def load_yaml(path: str) -> dict:
    """Load YAML file"""
